requirements.txt
web3==6.11.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
eth-account==0.10.0
eth-utils==2.3.1
websockets==12.0  # WebSocket 支持

# 可選但推薦
python-telegram-bot==20.7  # 如果要通知功能
colorama==0.4.6  # 美化終端輸出
playsound==1.2.2  # 音效通知（可選，Windows/Linux 需要）